+++ contrib/postgres.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS users
 (
-    user_id          INTEGER UNIQUE,
+    user_id          BIGINT UNIQUE,
     full_name        TEXT,
     username         TEXT,
     invite_code      TEXT,
@@ -9,21 +9,21 @@ CREATE TABLE IF NOT EXISTS users
 
 CREATE TABLE IF NOT EXISTS chat_messages
 (
-    user_id          INTEGER REFERENCES users(user_id),
-    chat_id          INTEGER,
+    user_id          BIGINT REFERENCES users(user_id),
+    chat_id          BIGINT,
     body             TEXT,
     created_at       timestamptz NOT NULL DEFAULT NOW()
 );
 
 CREATE TABLE IF NOT EXISTS wallets
 (
-    user_id          INTEGER UNIQUE REFERENCES users(user_id),
-    microdollars     INTEGER NOT NULL
+    user_id          BIGINT UNIQUE REFERENCES users(user_id),
+    microdollars     BIGINT NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS wallets_history
 (
-    user_id          INTEGER REFERENCES users(user_id),
-    microdollars     INTEGER NOT NULL,
+    user_id          BIGINT REFERENCES users(user_id),
+    microdollars     BIGINT NOT NULL,
     created_at       timestamptz NOT NULL DEFAULT NOW()
 );
